backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-decouple>=3.6
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
pydantic>=2.0.0
pyotp>=2.8.0
qrcode[pil]>=7.4.0
pandas>=2.2.0
openpyxl>=3.1.0
